First working version of federated Python
